Separated swap_chain_support_details struct
